pyproject.toml
[project]
name = "jetsetaccess"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.25",
    "firecrawl-py>=2.7.1",
    "selenium>=4.33.0",
    "streamlit>=1.45.1",
]
